[#182] refactor : Manage NAV_ITEM in src/constants
	여러 페이지에서 만들고 사용되던 NAV_ITEM를 src/constants 폴더에서 관리
