+++ src/constants/navigation/index.ts
@@ -1,0 +1,1 @@
+export { NAV_ITEMS } from "./navItems.tsx";
