--- src/pages/party/constants.tsx
@@ -1,10 +1,0 @@
-import { IconHome, IconNavigator, IconChat, IconPerson } from "jjan-icon";
-
-const NAV_ITEMS = [
-  { url: "/", label: "홈", icon: <IconHome /> },
-  { url: "/party-explore", label: "탐색하기", icon: <IconNavigator /> },
-  { url: "/chat-list", label: "채팅", icon: <IconChat /> },
-  { url: "/profile", label: "프로필", icon: <IconPerson /> },
-];
-
-export { NAV_ITEMS };
