+++ src/pages/profile/main/Main.tsx
@@ -9,9 +9,9 @@ import { Button } from "@/components/button";
 import { Flex } from "@/components/flex";
 import { Header } from "@/components/header";
 import { Layout } from "@/components/layout";
+import { NAV_ITEMS } from "@/constants/navigation";
 import { ProfileInfo } from "@/pages/components/profileInfo";
 import { ProfileOptionPanel } from "@/pages/components/profileOptions";
-import { NAV_ITEMS } from "@/pages/party/constants";
 
 const Main = () => {
   const navigate = useNavigate();
